refactor: Move ch. remainder theorem tests to test_maths
