+++ tests/test_maths.py
@@ -25,7 +25,8 @@ from algorithms.maths import (
     alice_private_key, alice_public_key, bob_private_key, bob_public_key, alice_shared_key, bob_shared_key, diffie_hellman_key_exchange,
     num_digits,
     alice_private_key, alice_public_key, bob_private_key, bob_public_key, alice_shared_key, bob_shared_key,
-    diffie_hellman_key_exchange, krishnamurthy_number
+    diffie_hellman_key_exchange, krishnamurthy_number,
+    chinese_remainder_theorem,
 )
 
 import unittest
@@ -465,6 +466,36 @@ class TestNumberOfDigits(unittest.TestCase):
         self.assertEqual(1,num_digits(-5))
         self.assertEqual(3,num_digits(-254))
 
+class TestChineseRemainderSolver(unittest.TestCase):
+    def test_k_three(self):
+        # Example which should give the answer 143 
+        # which is the smallest possible x that
+        # solves the system of equations
+        num = [3, 7, 10]
+        rem = [2, 3, 3]
+        self.assertEqual(chinese_remainder_theorem.solve_chinese_remainder(num, rem), 143)
+
+    def test_k_five(self):
+        # Example which should give the answer 3383 
+        # which is the smallest possible x that
+        # solves the system of equations
+        num = [3, 5, 7, 11, 26]
+        rem = [2, 3, 2, 6, 3]
+        self.assertEqual(chinese_remainder_theorem.solve_chinese_remainder(num, rem), 3383)
+
+    def test_exception_non_coprime(self):
+        # There should be an exception when all
+        # numbers in num are not pairwise coprime
+        num = [3, 7, 10, 14]
+        rem = [2, 3, 3, 1]
+        with self.assertRaises(Exception):
+            chinese_remainder_theorem.solve_chinese_remainder(num, rem)
+    
+    def test_empty_lists(self):
+        num = []
+        rem = []
+        with self.assertRaises(Exception):
+            chinese_remainder_theorem.solve_chinese_remainder(num, rem)
 
 if __name__ == "__main__":
     unittest.main()
